change object model
now we dont have a single type representing Fragment anymore but rather
two types: FragmentInner and Fragment. This allows us to get rid of some
data duplication (KeyParts) and some other wiredness.

There are still some hacks required to get the arg shouting going but
that is not nearly as bad as the stuff that was in place before.

also this paves the road to a brighter future where all non rsx widgets
decide themselves how to implement partial rebuild with regards to their
arguments and possibly a similiar mechanism for rsx widgets.