optimize KeyMap even further
Don't use a HashMap, but instead use the fact that the Keys are
linearly allocated from 0.

For each widget, store a array of children KeyPart's in a array indexd
by the Key and do a linear search of the children KeyPart's when looking
up the Key of a (parent key, KeyPart) tuple.

Makes circle_bench about 40% faster.