Implement subpasses
This add the notion of subpasses to the renderer. A subpass causes all
of its children to be rendered to its own framebuffer. This framebuffer
is then used as a input for two rendering passes
1. The resolve pass, a custom pass that is executed right
after all the children of the subpass have been rendered.
2. The finish pass, a custom pass that is executed as the last child of
the next higher subpass.
These subpasses can be nested and for a kind of stack with the first entry
in this stack being the actual framebuffer for the window.
They can be use to implement filters like blur and group opacity and
using depth tricks even fliters like backdrop blur (but only in a very
inefficient way).

Furthermore this also now sorts raw renderobjects properly in with
the rest (according to the z_index).

Finally this reworks the raw renderobjects to use secondary command
buffers (so they can properly bind their own vertex and index buffers
for example) and adds the vulkan context information (device, queues,
renderpass) to the WidgetContext to allow things like DescriptorSets to
be cached.