do not eagerly acquire new swapchain images
This seems to break MoltenVK and it is generally unclear if what we were doing here was legal.
Instead of this we should do proper frame pacing, but it seems like most of the vulkan extensions that would make this
better are not available yet. So for now just degrade the latency again.